consolidate styled components
