+++ frontend/alert/src/components/common/layout.js
@@ -11,11 +11,32 @@ export const Container = styled.div`
 
 export const Flex = styled.div`
     display: flex;
+    margin: ${props => props.margin};
+    padding: ${props => props.padding};
+    text-align: ${props => (props.center ? "center" : null)};
+    align-items: ${props => (props.valign ? "center" : null)};
+    justify-content: ${props => (props.halign ? "center" : props.spaceBetween ? "space-between" : null) /*eslint-disable-line*/};
     flex-direction: ${props => (props.col ? "column" : "row")};
-    align-items: ${props => props.align || "center"};
     flex-grow: ${props => props.grow || 0}
 `;
 
 export const Center = styled.div`
     text-align: center;
+`;
+
+export const GridItem = styled.div`
+    display: flex;
+    align-items: ${props => (props.valign ? "center" : null)};
+    justify-content: ${props => (props.halign ? "center" : null)};
+    grid-column: ${props => props.column};
+    grid-row: ${props => props.row};
+    background-color: ${props => (props.color ? props.color : "white")};
+    border-bottom: ${props => (props.border ? "1px solid #ececec" : null)};
+`;
+
+export const RightItem = styled.div`
+    height: 100%;
+    display: flex;
+    align-items: center;
+    margin-left: auto;
 `;
